refactor: make utils and value_specs more clear
