--- flexibleengine/util.go
@@ -1,116 +1,0 @@
-package flexibleengine
-
-import (
-	"fmt"
-
-	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
-	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
-	"github.com/huaweicloud/golangsdk"
-)
-
-// BuildRequest takes an opts struct and builds a request body for
-// Gophercloud to execute
-func BuildRequest(opts interface{}, parent string) (map[string]interface{}, error) {
-	b, err := golangsdk.BuildRequestBody(opts, "")
-	if err != nil {
-		return nil, err
-	}
-
-	b = AddValueSpecs(b)
-
-	return map[string]interface{}{parent: b}, nil
-}
-
-// CheckDeleted checks the error to see if it's a 404 (Not Found) and, if so,
-// sets the resource ID to the empty string instead of throwing an error.
-func CheckDeleted(d *schema.ResourceData, err error, msg string) error {
-	if _, ok := err.(golangsdk.ErrDefault404); ok {
-		d.SetId("")
-		return nil
-	}
-
-	return fmt.Errorf("%s: %s", msg, err)
-}
-
-// GetRegion returns the region that was specified in the resource. If a
-// region was not set, the provider-level region is checked. The provider-level
-// region can either be set by the region argument or by OS_REGION_NAME.
-func GetRegion(d *schema.ResourceData, config *Config) string {
-	if v, ok := d.GetOk("region"); ok {
-		return v.(string)
-	}
-
-	return config.Region
-}
-
-// AddValueSpecs expands the 'value_specs' object and removes 'value_specs'
-// from the reqeust body.
-func AddValueSpecs(body map[string]interface{}) map[string]interface{} {
-	if body["value_specs"] != nil {
-		for k, v := range body["value_specs"].(map[string]interface{}) {
-			body[k] = v
-		}
-		delete(body, "value_specs")
-	}
-
-	return body
-}
-
-// MapValueSpecs converts ResourceData into a map
-func MapValueSpecs(d *schema.ResourceData) map[string]string {
-	m := make(map[string]string)
-	for key, val := range d.Get("value_specs").(map[string]interface{}) {
-		m[key] = val.(string)
-	}
-	return m
-}
-
-// MapResourceProp converts ResourceData property into a map
-func MapResourceProp(d *schema.ResourceData, prop string) map[string]interface{} {
-	m := make(map[string]interface{})
-	for key, val := range d.Get(prop).(map[string]interface{}) {
-		m[key] = val.(string)
-	}
-	return m
-}
-
-func checkForRetryableError(err error) *resource.RetryError {
-	switch errCode := err.(type) {
-	case golangsdk.ErrDefault500:
-		return resource.RetryableError(err)
-	case golangsdk.ErrUnexpectedResponseCode:
-		switch errCode.Actual {
-		case 409, 503:
-			return resource.RetryableError(err)
-		default:
-			return resource.NonRetryableError(err)
-		}
-	default:
-		return resource.NonRetryableError(err)
-	}
-}
-
-func isResourceNotFound(err error) bool {
-	if err == nil {
-		return false
-	}
-	_, ok := err.(golangsdk.ErrDefault404)
-	_, ok1 := err.(golangsdk.ErrDefault404)
-	return ok || ok1
-}
-
-func hasFilledOpt(d *schema.ResourceData, param string) bool {
-	_, b := d.GetOkExists(param)
-	return b
-}
-
-// strSliceContains checks if a given string is contained in a slice
-// When anybody asks why Go needs generics, here you go.
-func strSliceContains(haystack []string, needle string) bool {
-	for _, s := range haystack {
-		if s == needle {
-			return true
-		}
-	}
-	return false
-}
